Fix layout for spotify card view
