+++ app/src/main/java/com/sregg/android/tv/spotifyPlayer/presenters/ArtistCardPresenter.java
@@ -3,7 +3,7 @@ package com.sregg.android.tv.spotifyPlayer.presenters;
 import android.support.v17.leanback.widget.Presenter;
 
 import com.sregg.android.tv.spotifyPlayer.SpotifyTvApplication;
-import com.sregg.android.tv.spotifyPlayer.views.NewSpotifyCardView;
+import com.sregg.android.tv.spotifyPlayer.views.SpotifyCardView;
 
 import java.net.URI;
 
@@ -22,7 +22,7 @@ public class ArtistCardPresenter extends AbsCardPresenter {
         super.onBindViewHolder(viewHolder, item);
 
         final CardViewHolder cardViewHolder = (CardViewHolder) viewHolder;
-        final NewSpotifyCardView imageCardView = cardViewHolder.getImageCardView();
+        final SpotifyCardView imageCardView = cardViewHolder.getImageCardView();
 
         ArtistSimple artistSimple = (ArtistSimple) item;
