+++ app/src/main/java/com/sregg/android/tv/spotifyPlayer/presenters/PlaylistSimpleCardPresenter.java
@@ -14,10 +14,8 @@
 
 package com.sregg.android.tv.spotifyPlayer.presenters;
 
-import android.support.v17.leanback.widget.ImageCardView;
-
 import com.sregg.android.tv.spotifyPlayer.R;
-import com.sregg.android.tv.spotifyPlayer.views.NewSpotifyCardView;
+import com.sregg.android.tv.spotifyPlayer.views.SpotifyCardView;
 
 import java.net.URI;
 
@@ -34,7 +32,7 @@ public class PlaylistSimpleCardPresenter extends AbsCardPresenter {
         super.onBindViewHolder(viewHolder, item);
 
         final CardViewHolder cardViewHolder = (CardViewHolder) viewHolder;
-        NewSpotifyCardView imageCardView = cardViewHolder.getImageCardView();
+        SpotifyCardView imageCardView = cardViewHolder.getImageCardView();
 
         final PlaylistSimple playlist = (PlaylistSimple) item;
