+++ app/src/main/java/com/sregg/android/tv/spotifyPlayer/presenters/TrackCardPresenter.java
@@ -17,7 +17,7 @@ package com.sregg.android.tv.spotifyPlayer.presenters;
 import android.support.v17.leanback.widget.Presenter;
 
 import com.sregg.android.tv.spotifyPlayer.utils.Utils;
-import com.sregg.android.tv.spotifyPlayer.views.NewSpotifyCardView;
+import com.sregg.android.tv.spotifyPlayer.views.SpotifyCardView;
 
 import java.net.URI;
 
@@ -34,7 +34,7 @@ public class TrackCardPresenter extends AbsCardPresenter {
         super.onBindViewHolder(viewHolder, item);
 
         final CardViewHolder cardViewHolder = (CardViewHolder) viewHolder;
-        final NewSpotifyCardView imageCardView = cardViewHolder.getImageCardView();
+        final SpotifyCardView imageCardView = cardViewHolder.getImageCardView();
 
         Track track = (Track) item;
