+++ app/src/main/res/layout/spotify_card_view.xml
@@ -2,21 +2,38 @@
 <merge xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:lb="http://schemas.android.com/apk/res-auto">
 
-    <android.support.v17.leanback.widget.NonOverlappingFrameLayout
+    <FrameLayout
+        android:id="@+id/main_container"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content">
 
         <ImageView
             android:id="@+id/image_view_main"
             android:layout_width="match_parent"
-            android:layout_height="match_parent" />
+            android:layout_height="match_parent"/>
 
-        <com.sregg.android.tv.spotifyPlayer.views.NowPlayingIndicatorView
-            android:id="@+id/now_playing_indicator_view"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent" />
+        <FrameLayout
+            android:id="@+id/now_playing_container"
+            android:layout_width="@dimen/card_image_view_size"
+            android:layout_height="@dimen/card_image_view_size">
 
-    </android.support.v17.leanback.widget.NonOverlappingFrameLayout>
+            <ImageView
+                android:id="@+id/card_badge"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center"
+                android:scaleType="center"
+                android:src="@drawable/ic_card_badge_bg"/>
+
+            <com.sregg.android.tv.spotifyPlayer.views.NowPlayingIndicatorView
+                android:id="@+id/now_playing_indicator_view"
+                android:layout_width="35dp"
+                android:layout_height="35dp"
+                android:layout_gravity="center"/>
+
+        </FrameLayout>
+
+    </FrameLayout>
 
     <android.support.v17.leanback.widget.NonOverlappingFrameLayout
         android:layout_width="match_parent"
